Remove next connect and useless stuff.
